Fixed bug where clicking on viewitems was being redirected to dashboard by adding HttpSession as parameters for functions(Httpsession was null). Also changed the display style for item page's createdBy and updatedBY as usernames rather than ids

+++ src/main/java/com/gmi/learn/dao/impl/FoodDaoImpl.java
@@ -49,8 +49,8 @@ public class FoodDaoImpl implements FoodDao {
                             rs.getString("name"),
                             rs.getLong("price"),
                             rs.getString("image"),
-                            rs.getInt("lastUpdated"),
-                            rs.getInt("createdBy"),
+                            rs.getLong("lastUpdated"),
+                            rs.getLong("createdBy"),
                             rs.getString("updatedBy"),
                             rs.getBoolean("isDeleted")
                     )
@@ -73,8 +73,8 @@ public class FoodDaoImpl implements FoodDao {
                         rs.getString("name"),
                         rs.getLong("price"),
                         rs.getString("image"),
-                        rs.getInt("lastUpdated"),
-                        rs.getInt("createdBy"),
+                        rs.getLong("lastUpdated"),
+                        rs.getLong("createdBy"),
                         rs.getString("updatedBy"),
                         rs.getBoolean("isDeleted")
                 )
